+++ core-ng-search/src/test/java/core/framework/search/ElasticSearchMigrationTest.java
@@ -1,0 +1,26 @@
+package core.framework.search;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+/**
+ * @author neo
+ */
+class ElasticSearchMigrationTest {
+    private ElasticSearchMigration migration;
+
+    @BeforeEach
+    void createElasticSearchMigration() {
+        migration = new ElasticSearchMigration("migration-test/sys.properties");
+    }
+
+    @Test
+    void migrate() {
+        assertThatThrownBy(() -> migration.migrate(search -> {
+            throw new RuntimeException("migration error");
+        })).isInstanceOf(RuntimeException.class)
+           .hasMessageContaining("migration");
+    }
+}